Implement Uploader
Did not use apache module.
Although it takes in a LocationJSON, it’s basically an abstract
uploader for JSON’s.
Might need to fix LocationJSON so that it stores them in a map.
Did not check for getActiveNetworkInfo() or isConnected(), might be
better to check before call, if not easy fix