Fixed constant names and changed ResultCallable to Object
Changed result callable to object in order to allow a setter for it in
Uploader.

Also, I thought that params was just going to be an array of JSON
objects, this allows you to store different types that aren't
necessarily string types. What is the benefit of the other way?